healing-engine/requirements.txt
numpy<2.0.0
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
jinja2==3.1.3
aiosmtplib==3.0.1
chromadb==0.4.22
anthropic==0.43.0
openai==1.63.2
google-generativeai==0.8.4
